requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langsmith
docarray
